don't allow user to delete all service providers,  make sure at least one remains [#69839352]

+++ app/controllers/catalog_manager/identities_controller.rb
@@ -65,8 +65,17 @@ class CatalogManager::IdentitiesController < CatalogManager::AppController
 
     if rel_type == 'service_provider_organizational_unit'
       service_provider = ServiceProvider.find params["relationship"]
-      service_provider.destroy
+      @service_provider_error_message = nil
+      
+      if oe.all_service_providers.size > 1 # we need to have more than just this one service provider in the tree in order to delete
+        service_provider.destroy
+        oe.reload
+      else
+        @service_provider_error_message = I18n.t("organization_form.service_provider_required_message") 
+      end
+      
       render :partial => 'catalog_manager/shared/service_providers', :locals => {:entity => oe}
+
     elsif rel_type == 'super_user_organizational_unit'
       super_user = SuperUser.find params["relationship"]
       super_user.destroy
